Split listing/git helpers into focused files
