+++ internal/ui/views/listing/listing.go
@@ -78,11 +78,6 @@ type Model struct {
 	notifier         *notifications.Notifier
 }
 
-type ActivityFinalizeResult struct {
-	Completed bool
-	Info      InfoMessageResult
-}
-
 func New(repos []domain.Repository) *Model {
 	return NewWithNotifier(repos, nil)
 }
@@ -386,19 +381,6 @@ func (m *Model) applyPullRequestWatchlist(tracked []pullrequests.Snapshot, seed 
 	}
 
 	m.prCoordinator.Sync(tracked, pullrequests.ApplyOptions{Seed: seed}, m.notifier)
-}
-
-func (m *Model) finalizeRepoActivity(index int, next domain.Repository, complete func(activity domain.Activity) ActivityFinalizeResult) {
-	m.applyRepoUpdate(index, next, func(repo *domain.Repository, activity domain.Activity) {
-		result := complete(activity)
-		if !result.Completed {
-			return
-		}
-		if result.Info.OK {
-			m.storeRecentActivityInfo(repo.Path, result.Info.Message)
-		}
-		repo.Activity = &domain.IdleActivity{}
-	})
 }
 
 func (m *Model) View() string {
